# cmd/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Esta API Json Rest actua como proxy SOAP a los servicios web de ARCA.",
        "title": "API proxy a los webservices de ARCA",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Sebastian Hogas",
            "email": "sehogas@gmail.com"
        },
        "version": "1.0"
    },
    "paths": {
        "/coem/AnularCOEM": {
            "delete": {
                "description": "Método a través del cual se anula una COEM. La COEM debe estar en estado en CURSO o REGISTRADA.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Comunicación de Embarque"
                ],
                "summary": "Anular COEM",
                "parameters": [
                    {
                        "type": "string",
                        "description": "API Key de acceso",
                        "name": "x-api-key",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "AnularCOEMRequest",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/wscoem.AnularCOEMRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/wscoem.AnularEmbarqueRta"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/coem/AnularCaratula": {
            "delete": {
                "description": "Método que permite eliminar una carátula. Si se encuentran COEMs en estado Presentada o Autorizada se detiene la ejecución.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Comunicación de Embarque"
                ],
                "summary": "Anular Carátula",
                "parameters": [
                    {
                        "type": "string",
                        "description": "API Key de acceso",
                        "name": "x-api-key",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "AnularCaratulaRequest",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/wscoem.AnularCaratulaRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/wscoem.AnularEmbarqueRta"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/coem/CerrarCOEM": {
            "post": {
                "description": "Método a través del cual se cierra la carga de una COEM asociada a una Carátula, permitiendo que el operador Aduanero pueda trabajar con ella. La COEM debe estar en estado CUR.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Comunicación de Embarque"
                ],
                "summary": "Cerrar COEM",
                "parameters": [
                    {
                        "type": "string",
                        "description": "API Key de acceso",
                        "name": "x-api-key",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "CerrarCOEMRequest",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/wscoem.CerrarCOEMRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/wscoem.RegistrarEmbarqueRta"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/coem/Dummy": {
            "get": {
                "description": "Visualizar el estado del servicio web, del servicio de autenticación y de la base de datos de ARCA",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Comunicación de Embarque"
                ],
                "summary": "Muestra el estado del servicio",
                "parameters": [
                    {
                        "type": "string",
                        "description": "API Key de acceso",
                        "name": "x-api-key",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/wscoem.ResultadoEjecucionDummy"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/coem/RectificarCOEM": {
            "put": {
                "description": "Método a través del cual se modifican los valores de una COEM. La COEM debe estar en curso o registrada. No se habilita rectificar una COEM presentada o autorizada. No debe existir solicitud de cierre de carga iniciada o aprobada.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Comunicación de Embarque"
                ],
                "summary": "Rectificar COEM",
                "parameters": [
                    {
                        "type": "string",
                        "description": "API Key de acceso",
                        "name": "x-api-key",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "RectificarCOEMRequest",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/wscoem.RectificarCOEMRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/wscoem.RectificarEmbarqueRta"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/coem/RectificarCaratula": {
            "put": {
                "description": "Permite rectificar los datos de una Carátula previamente creada con el método RegistrarCaratula. Entre las restricciones, no se permitirá cargar datos idénticos a otra Carátula existente, ni modificar aquella carátula que tenga COEMs asociados.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Comunicación de Embarque"
                ],
                "summary": "Rectificar Carátula",
                "parameters": [
                    {
                        "type": "string",
                        "description": "API Key de acceso",
                        "name": "x-api-key",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "RectificarCaratulaRequest",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/wscoem.RectificarCaratulaRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/wscoem.RectificarEmbarqueRta"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "dto.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "wscoem.AnularCOEMRequest": {
            "type": "object",
            "properties": {
                "IdentificadorCOEM": {
                    "type": "string"
                },
                "IdentificadorCaratula": {
                    "type": "string"
                }
            }
        },
        "wscoem.AnularCaratulaRequest": {
            "type": "object",
            "properties": {
                "IdentificadorCaratula": {
                    "type": "string"
                }
            }
        },
        "wscoem.AnularEmbarqueRta": {
            "type": "object",
            "properties": {
                "ListaErrores": {
                    "$ref": "#/definitions/wscoem.ArrayOfDgaMensajeSistemaReducido"
                },
                "Server": {
                    "type": "string"
                },
                "TimeStamp": {
                    "type": "string"
                }
            }
        },
...
        "wscoem.CerrarCOEMRequest": {
            "type": "object",
            "properties": {
                "IdentificadorCOEM": {
                    "type": "string"
                },
                "IdentificadorCaratula": {
                    "type": "string"
                }
            }
        },
...
        "wscoem.RegistrarEmbarqueRta": {
            "type": "object",
            "properties": {
                "ListaErrores": {
                    "$ref": "#/definitions/wscoem.ArrayOfDgaMensajeSistemaReducido"
                },
                "Server": {
                    "type": "string"
                },
                "TimeStamp": {
                    "type": "string"
                }
            }
        },
        "wscoem.ResultadoEjecucionDummy": {
            "type": "object",
            "properties": {
                "AppServer": {
                    "type": "string"
                },
                "AuthServer": {
                    "type": "string"
                },
                "DbServer": {
                    "type": "string"
                }
            }
        },
...
    }
}